requirements.txt
sqlmodel==0.0.25
alembic==1.16.5
psycopg2-binary==2.9.10
asyncpg==0.30.0
SQLAlchemy-Utils==0.42.1


grpcio==1.75.0
grpcio-tools==1.75.0

pytest==9.0.2
